requirements.txt
djangorestframework==3.13.1
Faker==13.15.1
drf-url-filters==0.3.15
DjangoFilterBackend==22.1